Extract docgen tasks
